requirements.txt
argbind
descript-audiotools
numpy
primePy
torch
torchaudio
tqdm
tensorboard
jupyterlab
